pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "autohedge"
version = "0.2.0"
description = "AutoHedge - Automated Trading Insights System using Crew AI"
license = "MIT"
authors = ["Kye Gomez <kye@apac.ai>"]
homepage = "https://github.com/The-Swarm-Corporation/AutoHedge"
readme = "README.md"
repository = "https://github.com/The-Swarm-Corporation/AutoHedge"
keywords = ["trading", "ai", "crewai", "hedge-fund", "automation", "insights"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Office/Business :: Financial :: Investment",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

packages = [
    { include = "autohedge" }
]

[tool.poetry.dependencies]
python = "^3.9"
crewai = "*"
yfinance = "*"
pydantic = "^2.0"
loguru = "*"
python-dotenv = "*"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
black = "^23.1.0"
pytest = "^7.0"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
